HITW: completely overhauled the BuildLoader.kt (again lol). it was very, very painful and long.
main thing to take away from -
the schematic loading methods are more 'component-based' - you can take a schematic and modify it more freely with more options.

added as well a method that takes the schematic and mirrors it in a way that it keeps its location and the direction it is facing. [took way too long ;-;]

this overhaul makes it so that giving extra features in the future is more easy and convenient