HITW: Added WallSpawnerModes - modes dictate the list of 'behaviors' that walls can have, and only have them at that mode.
 A behavior, for example, is - should a wall be spawning in the same direction as the previous added wall or not.

Currently, there are 2 modes, which can be selected to be played, and also an "Alternated" keyword that switches between  all available modes throughout the game, instead of sticking to 1 mode.

the current modes are:
- WALL_CHAINER - walls rapidly come from one direction as one tight group. occusionaly the spawn direction is changed up.
- WALLS_FROM_ALL_DIRECTIONS - walls will spawn in groups from 2 to 4 *at once*, from all directions. one of the walls will not stop at all, while the others will, in attempt to fool the player. these are the Psych walls that were implemented earlier.

More modes to come!!!

alongside the modes accompanies a WallSpawner State Machine that handles all states that the wall spawner can have. The Modes dictate which States can be chosen from.