HITW: reworked the latest wall spawning mode added so that the directions of the duo walls may sometimes change. also, now, when there seems like the real wall direction is gonna switch from the duo directions, it now only has half a chance of doing so - otherwise the real wall direction stays the same, and there's only a wave of psych walls from both directions for a small amount of time
on /start brought back the wall spawning mode parameter, this time as an optional parameter.

also if you set the mode to "alternating" while the game is alive, it'll now actually activate.

the wall spawning mode now resets to null in between running games, so each time you start a game u will need to set ahead of time the mode you want to be in.