Project now tracks its "version"
Format: [Cur Minigame num being worked on].[Big Addition added to minigame index].[small addition/bug fixes index]

for example, 3.3.1 means that this commit was for the minigame Hitw, which has had 3 major features added, and has (now) got 1 small addition/ bug fixes.

Version is now 4.0.0 , which resembles the want of a new minigame in the works. Additions to it will be shown via this tracker.