Reworked internally how the Lifecycle methods in MinigameSkeleton.kt work -
each Method (start/pause/resume/end) now has its "skeleton" variant - this variant is inlined and has all the code that those methods previously had.

The proper methods need to be ovveridden with those inline variants in their method.

This refactoring makes it so that, because the "skeleton" varaiants are inlined, early returns from them *exit* the outer method, which makes for neat and easy exits.

[NOT TESTED]